% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_catalogue.R
\name{expand_catalogue}
\alias{expand_catalogue}
\title{Adding Columns to the DHS Catalog}
\usage{
expand_catalogue(x)
}
\arguments{
\item{x}{a data frame catalog has outputed by \code{\link[lodown]{get_catalog}}.}
}
\value{
a data frame catalog has outputed by \code{\link[lodown]{get_catalog}},
with three additional columns: \code{file_type}, \code{version_number} and
\code{file_format}. See example in the documentation of
\code{\link{with_profile}}.
}
\description{
\code{expand_catalogue} adds file type, version number and file format
information to the the catalog as generated by the
\code{\link[lodown]{get_catalog}} function.
}
