% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catalog2.R
\name{get_catalog2}
\alias{get_catalog2}
\title{Download the Catalog of a Profile's Project}
\usage{
get_catalog2(profile)
}
\arguments{
\item{profile}{named character vector with fields \code{data_name},
\code{your_email}, \code{your_password} and \code{your_project}.}
}
\description{
\code{get_catalog2} is a wrapper around the \code{\link[lodown]{get_catalog}}
function. Before downloading the catalog corresponding to the project of the
specified profile, the function checks in the global environment whether there
is already a catalog corresponding to the same project. If so, the function
stops returning the names of the catalogs of the global environment that
correspond to this project. If not, the function proceed to download the
catalog and returns it with an "project" attribute that contains the name of
the project of the inputed profile. This avoid downloading the same catalog
twice.
}
