% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dhs.R
\name{load_dhs}
\alias{load_dhs}
\title{Load data from DHS}
\usage{
load_dhs(profile, ..., output = getwd())
}
\arguments{
\item{profile}{named character vector with fields \code{data_name},
\code{your_email}, \code{your_password} and \code{your_project}.}

\item{...}{logical dplyr::filter predicates defined in terms of the
variables in the catalog.}

\item{output}{name of the directory in which the files downloaded from the
DHS server should be saved.}
}
\value{
This function does not return any thing and is used for its side
effects, which are (i) creating the catalog corresponding to the inputed
profile's project and saving it into the global environment, (ii) downloading
the data to disk from the DHS server, and (iii) loading the downloaded data
from disk to the global environment.
}
\description{
\code{load_dhs} (i) makes the catalog corresponding to the inputed profile's
project, (ii) optionally filters it according to conditions on countries,
years, and type of data, (iii) download the data file from the DHS server to
disk and (iv) load these data files to the global environment.
}
\details{
Of note, this function is supposed to be efficient in doing what it does. In
particular, both the creation of the catalog and the downloading of data from
the DHS server can be time consuming in that they query the DHS server.
Furthermore, the loading of the data from disk to the global environment may
also take time depending on the size and number of data files to load. Thus,
each of these actions are preceded by checking whether they have previously
done. If yes, they are not done again. If no, it also makes sure that it won't
have to be done for a second call of the function. For example, if a
catalog corresponding to the inputed profile's project already exist in the
global environment, it won't be created again and this already existing
catalog will be used instead. If such a catalog does not exist, then it will
be created and used inside the function but also it will be saved in the
global environment for potential future calls of the function on the same
profile.
}
\examples{
## Run the following call twice and see the differences in execution times:
\dontrun{
system.time(load_dhs(my_profile, year == 2015, country == "Tanzania", file_type == "KR"))
system.time(load_dhs(my_profile, year == 2015, country == "Tanzania", file_type == "KR"))
}

## Try deleting the data object and rerun:
\dontrun{
rm("TZKR7AFL")
system.time(load_dhs(my_profile, year == 2015, country == "Tanzania", file_type == "KR"))
}

## Try deleting the data file and rerun:
\dontrun{
file.remove("TZKR7AFL.rds")
system.time(load_dhs(my_profile, year == 2015, country == "Tanzania", file_type == "KR"))
}

## Try deleting the catalog and rerun:
\dontrun{
rm("my_profile_catalog")
system.time(load_dhs(my_profile, year == 2015, country == "Tanzania", file_type == "KR"))
}

}
