% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lodown2.R
\name{lodown2}
\alias{lodown2}
\title{Locally Download Publicly-available data}
\usage{
lodown2(catalog = NULL)
}
\arguments{
\item{catalog}{\code{data.frame} detailing available microdata extracts and
metadata, see the documentation of \code{\link[lodown]{lodown}}.}

\item{...}{additional arguments passed to \code{\link[lodown]{lodown}}.}
}
\description{
\code{lodown2} is a wrapper around the \code{\link[lodown]{lodown}} function
that checks the existance of files on disk before trying to download them. It
will thus download only the files that are not already on disk.
}
\details{
See examples in the \code{\link{with_profile}} documentation.
}
