% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mics.R
\docType{package}
\name{mics-package}
\alias{mics}
\alias{mics-package}
\title{mics: analysing multiple indicator cluster surveys}
\description{
mics provides a collection of functions to analyse multiple indicator cluster
surveys such as the UNICEF MICS or USAID DHS.
}
\details{
To learn more about mics, start with the vignettes:
`browseVignettes(package = "mics")`
}
\author{
\strong{Maintainer}: Marc Choisy \email{marc.choisy@ird.fr} (0000-0002-5187-6390) [copyright holder]

Authors:
\itemize{
  \item Sonia Lewycka \email{slewycka@oucru.org} (0000-0002-5923-9468) [copyright holder]
  \item Lucie Contamin \email{contamin.lucie@gmail.com} (0000-0001-5797-1279) [copyright holder]
}

}
